% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{check_not_overlap}
\alias{check_not_overlap}
\title{Check not overlap}
\usage{
check_not_overlap(starttime1, stoptime1, starttime2, stoptime2)
}
\arguments{
\item{starttime1}{A string or time of the form "H:M pm", e.g. "9:00 pm"}

\item{stoptime1}{A string or time of the form "H:M pm", e.g. "9:00 pm"}

\item{starttime2}{A string or time of the form "H:M pm", e.g. "9:00 pm"}

\item{stoptime2}{A string or time of the form "H:M pm", e.g. "9:00 pm"}
}
\value{

}
\description{
Checks for a schedule conflict given starttimes and endtimes.
}
