% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{htmltxt_to_table_schedule}
\alias{htmltxt_to_table_schedule}
\title{Read an html table stored in filename following a specific format to a tibble, with cleaning.}
\usage{
htmltxt_to_table_schedule(filename, by_day = FALSE)
}
\arguments{
\item{filename}{A text file containing an html table following a specific format.}

\item{by_day}{If TRUE, every day is a unique row.}
}
\value{

}
\description{
Some classes have extra times for a given table row. These appear with an extra Day not separated by a comma.
For example, "Mon,Wed,FriThu" has an extra meeting time for Thursday. These entries also have two Start and Stop times.
This function separates this into two rows, e.g. one for "Mon,Wed,Fri" and one for "Thu".
}
