% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{trim_off_number}
\alias{trim_off_number}
\title{Trim off last number
Remove the last number of a course title. Course titles have a form like "CAS MA415 A1", and we would like to remove any numbers.}
\usage{
trim_off_number(A1)
}
\arguments{
\item{A1}{A string that contains numbers to be removed.}
}
\value{

}
\description{
Trim off last number
Remove the last number of a course title. Course titles have a form like "CAS MA415 A1", and we would like to remove any numbers.
}
